requirements.txt
requests>=2.31.0, < 3.0
google-api-python-client>=2.127.0 < 3.0
google-auth-oauthlib>=1.2.0 < 2.0